+++ src/lib/framework_constants.ts
@@ -3,3 +3,8 @@ export const NEXTJS_CONFIG_FILES = [
   "next.config.mjs",
   "next.config.ts",
 ];
+
+export function isNextJsProject(files: string[] | undefined): boolean {
+  if (!files) return false;
+  return files.some((file) => NEXTJS_CONFIG_FILES.includes(file));
+}
